Make FormData.entries polyfill return iterator
The polyfill can return an iterator by manually constructing an object
with next() function, as opposed to using a generator function which is
not supported by the current Babel configuration.
