+++ src/polyfills.js
@@ -26,9 +26,7 @@ import {install as installMathSign} from './polyfills/math-sign';
 import {install as installObjectAssign} from './polyfills/object-assign';
 import {install as installPromise} from './polyfills/promise';
 import {install as installArrayIncludes} from './polyfills/array-includes';
-import {
-  install as installFormDataEntryArray,
-} from './polyfills/formdata-entryarray';
+import {install as installFormDataEntries} from './polyfills/formdata-entries';
 import {getMode} from './mode';
 
 /**
@@ -47,4 +45,4 @@ installObjectAssign(self);
 installPromise(self);
 installDocContains(self);
 installArrayIncludes(self);
-installFormDataEntryArray(self);
+installFormDataEntries(self);
